Fixed up side bar a bit.
We removed the HEADER declaration from the
HOME like, so that no longer displays as a big
annoying blue box. Also, by re-wrapping
some of the div in the default.html
file around the page links, we can
stack the social links side by side on
a line, as opposed to being drawn out
down the side.