Linking between research and academic pages.
Currently all of the stuff related to my actual
research is put on the academics page, while
the stuff described on the research page is mostly
just high level. I put a link on the research page
back to the academics page.

Also, changed the markdown header format
on the academics page.