Fixing up some of the site links to work with new file structure.
In particular, we now have the _data
folder, in which we can store various
YAML files. One such file contains a list
of all the social links I have. This
way I can add/subtract using that list, without
having to screw around and change the links
in the sidebar.html file itself.
(Of course, if I remove a link, I had better
remove the reference to it!)

Additionally, this _data folder provides
us with a way to centralize links, so that
I can easily link to the same site(s) just
by reference.