Writing a layout for a talk page.
I wanted to start putting each of my talks on their own page so
that I could share the slides, abstract, description, etc.
without cluttering up my academics page. This layout is essentially
the same as the default one, except we allow for a "location"
specification in the YAML front matter.