Making talk pages.
To make a talk page, I simply took the markdown I used for blog
posts and modified the YAML front matter some. In particular,
we introduce a "type" to the post. Currently, "post" or "talk"
is supported. This way the musings page can display posts, and the
academics page can display talks (eventually posters), and
we can put all the files inside the _posts/ directory.

Essentially, we moved all the stuff about talks onto their own
pages, where I put a description, an abstract, a copy of the
slide deck, and a screencast, if possible.