Adjusting font size and bottom margin.
Setting the font size for the content to be 1em. This
explicitly declares the size and, by using the em unit,
makes it possible for people to resize the font using
their browser settings.

Made the bottom margin a bit smaller to reduce overall page
space. Question: How to reduce side margins while avoiding huge
wide paragraphs of text?